Provide core-image-readonly-rootfs-overlay-initramfs as an alias
The image currently exports IMAGE_BASENAME as
"core-image-readonly-rootfs-overlay-initramfs" which is then used as
the deployed image name.

This is an issue in situations where the provided image name is used for
searching for the corresponding artifact in ${DEPLOY} directory.

Example on such usage is the INITRAMFS_IMAGE variable in the kernel-fitimage class.

Therefor, provide an alias that matches the artifact image.

Signed-off-by: Marcus Folkesson <marcus.folkesson@gmail.com>